use locl for i18n extractions
Since "ng extract-i18n" only scans templates, but not TypeScript code,
another extractor is needed.  locl fills this role.

Switch output format from XLF to JSON, because it is smaller and can
be loaded at runtime (which is exactly what we want and what regular
Angular also can't do by itself).

locl needs the app to be built with all locales at once, so we don't
limit the build to a single locale one any more (as it was needed
before by the Angular tooling).